Revert "[NUI] Add BindableProperties to all public Properties"
This reverts commit 232d139ae0494164d244f50cf6012ee34e2029e8.
