+++ src/Tizen.NUI/src/public/BaseComponents/AnimatedImageView.cs
@@ -29,7 +29,7 @@ namespace Tizen.NUI.BaseComponents
     /// </summary>
     // This will be public opened after ACR done. (Before ACR, need to be hidden as Inhouse API)
     [EditorBrowsable(EditorBrowsableState.Never)]
-    public partial class AnimatedImageView : ImageView
+    public class AnimatedImageView : ImageView
     {
         #region Constructor, Destructor, Dispose
         /// <summary>
@@ -72,19 +72,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return GetValue(ResourceUrlProperty) as string;
-            }
-            set
-            {
-                SetValue(ResourceUrlProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private string InternalResourceUrl
-        {
-            get
-            {
                 return url;
             }
             set
@@ -117,19 +104,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return (int)GetValue(BatchSizeProperty);
-            }
-            set
-            {
-                SetValue(BatchSizeProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalBatchSize
-        {
-            get
-            {
                 return batchSize;
             }
             set
@@ -150,19 +124,6 @@ namespace Tizen.NUI.BaseComponents
         // This will be public opened after ACR done. (Before ACR, need to be hidden as Inhouse API)
         [EditorBrowsable(EditorBrowsableState.Never)]
         public int CacheSize
-        {
-            get
-            {
-                return (int)GetValue(CacheSizeProperty);
-            }
-            set
-            {
-                SetValue(CacheSizeProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalCacheSize
         {
             get
             {
@@ -188,19 +149,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return (int)GetValue(FrameDelayProperty);
-            }
-            set
-            {
-                SetValue(FrameDelayProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalFrameDelay
-        {
-            get
-            {
                 return frameDelay;
             }
             set
@@ -219,19 +167,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return (int)GetValue(LoopCountProperty);
-            }
-            set
-            {
-                SetValue(LoopCountProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalLoopCount
-        {
-            get
-            {
                 return loopCount;
             }
             set
@@ -246,19 +181,6 @@ namespace Tizen.NUI.BaseComponents
         /// </summary>
         [EditorBrowsable(EditorBrowsableState.Never)]
         public StopBehaviorType StopBehavior
-        {
-            get
-            {
-                return (StopBehaviorType)GetValue(StopBehaviorProperty);
-            }
-            set
-            {
-                SetValue(StopBehaviorProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private StopBehaviorType InternalStopBehavior
         {
             set
             {
@@ -304,19 +226,6 @@ namespace Tizen.NUI.BaseComponents
         /// </remarks>
         [EditorBrowsable(EditorBrowsableState.Never)]
         public int CurrentFrame
-        {
-            get
-            {
-                return (int)GetValue(CurrentFrameProperty);
-            }
-            set
-            {
-                SetValue(CurrentFrameProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalCurrentFrame
         {
             set
             {
