+++ src/Tizen.NUI/src/public/BaseComponents/AnimatedVectorImageView.cs
@@ -28,7 +28,7 @@ namespace Tizen.NUI.BaseComponents
     /// AnimatedVectorImageView is a class for displaying a vector resource.
     /// </summary>
     [EditorBrowsable(EditorBrowsableState.Never)]
-    public partial class AnimatedVectorImageView : LottieAnimationView
+    public class AnimatedVectorImageView : LottieAnimationView
     {
         #region Constructor, Destructor, Dispose
         /// <summary>
@@ -82,19 +82,6 @@ namespace Tizen.NUI.BaseComponents
         [EditorBrowsable(EditorBrowsableState.Never)]
         public string ResourceURL
         {
-            get
-            {
-                return GetValue(ResourceURLProperty) as string;
-            }
-            set
-            {
-                SetValue(ResourceURLProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private string InternalResourceURL
-        {
             set
             {
                 tlog.Fatal(tag, $"[AnimatedVectorImageView START[ [{GetId()}] ResourceURL SET");
@@ -120,19 +107,6 @@ namespace Tizen.NUI.BaseComponents
         [EditorBrowsable(EditorBrowsableState.Never)]
         public new string ResourceUrl
         {
-            get
-            {
-                return GetValue(ResourceUrlProperty) as string;
-            }
-            set
-            {
-                SetValue(ResourceUrlProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private string InternalResourceUrl
-        {
             set
             {
                 tlog.Fatal(tag, $"[AnimatedVectorImageView START[ [{GetId()}] ResourceUrl SET");
@@ -156,19 +130,6 @@ namespace Tizen.NUI.BaseComponents
         /// </summary>
         [EditorBrowsable(EditorBrowsableState.Never)]
         public int RepeatCount
-        {
-            get
-            {
-                return (int)GetValue(RepeatCountProperty);
-            }
-            set
-            {
-                SetValue(RepeatCountProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalRepeatCount
         {
             set
             {
@@ -197,19 +158,6 @@ namespace Tizen.NUI.BaseComponents
         /// <returns> Returns user set value for the current frame. Cannot provide actual playing current frame. </returns>
         [EditorBrowsable(EditorBrowsableState.Never)]
         public new int CurrentFrame
-        {
-            get
-            {
-                return (int)GetValue(CurrentFrameProperty);
-            }
-            set
-            {
-                SetValue(CurrentFrameProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalCurrentFrame
         {
             set
             {
@@ -245,19 +193,6 @@ namespace Tizen.NUI.BaseComponents
         /// </summary>
         [EditorBrowsable(EditorBrowsableState.Never)]
         public RepeatModes RepeatMode
-        {
-            get
-            {
-                return (RepeatModes)GetValue(RepeatModeProperty);
-            }
-            set
-            {
-                SetValue(RepeatModeProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private RepeatModes InternalRepeatMode
         {
             set
             {
