+++ src/Tizen.NUI/src/public/BaseComponents/LottieAnimationView.cs
@@ -31,7 +31,7 @@ namespace Tizen.NUI.BaseComponents
     /// LottieAnimationView renders an animated vector image (Lottie file).
     /// </summary>
     /// <since_tizen> 7 </since_tizen>
-    public partial class LottieAnimationView : ImageView
+    public class LottieAnimationView : ImageView
     {
         #region Constructor, Destructor, Dispose
         /// <summary>
@@ -106,19 +106,6 @@ namespace Tizen.NUI.BaseComponents
         /// </summary>
         /// <since_tizen> 7 </since_tizen>
         public string URL
-        {
-            get
-            {
-                return GetValue(URLProperty) as string;
-            }
-            set
-            {
-                SetValue(URLProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private string InternalURL
         {
             set
             {
@@ -247,19 +234,6 @@ namespace Tizen.NUI.BaseComponents
         /// <since_tizen> 7 </since_tizen>
         public int CurrentFrame
         {
-            get
-            {
-                return (int)GetValue(CurrentFrameProperty);
-            }
-            set
-            {
-                SetValue(CurrentFrameProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalCurrentFrame
-        {
             set
             {
                 currentStates.frame = value;
@@ -292,19 +266,6 @@ namespace Tizen.NUI.BaseComponents
         /// </summary>
         /// <since_tizen> 7 </since_tizen>
         public LoopingModeType LoopingMode
-        {
-            get
-            {
-                return (LoopingModeType)GetValue(LoopingModeProperty);
-            }
-            set
-            {
-                SetValue(LoopingModeProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private LoopingModeType InternalLoopingMode
         {
             set
             {
@@ -364,19 +325,6 @@ namespace Tizen.NUI.BaseComponents
         /// <since_tizen> 7 </since_tizen>
         public int LoopCount
         {
-            get
-            {
-                return (int)GetValue(LoopCountProperty);
-            }
-            set
-            {
-                SetValue(LoopCountProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private int InternalLoopCount
-        {
             set
             {
                 currentStates.changed = true;
@@ -418,19 +366,6 @@ namespace Tizen.NUI.BaseComponents
         /// </summary>
         /// <since_tizen> 7 </since_tizen>
         public StopBehaviorType StopBehavior
-        {
-            get
-            {
-                return (StopBehaviorType)GetValue(StopBehaviorProperty);
-            }
-            set
-            {
-                SetValue(StopBehaviorProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private StopBehaviorType InternalStopBehavior
         {
             set
             {
@@ -478,19 +413,6 @@ namespace Tizen.NUI.BaseComponents
         /// </remarks>
         [EditorBrowsable(EditorBrowsableState.Never)]
         public bool RedrawInScalingDown
-        {
-            get
-            {
-                return (bool)GetValue(RedrawInScalingDownProperty);
-            }
-            set
-            {
-                SetValue(RedrawInScalingDownProperty, value);
-                NotifyPropertyChanged();
-            }
-        }
-
-        private bool InternalRedrawInScalingDown
         {
             set
             {
