+++ src/Tizen.NUI/src/public/BaseComponents/TextEditor.cs
@@ -99,18 +99,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return GetValue(TranslatableTextProperty) as string;
-            }
-            set
-            {
-                SetValue(TranslatableTextProperty, value);
-            }
-        }
-
-        private string InternalTranslatableText
-        {
-            get
-            {
                 return textEditorTextSid;
             }
             set
@@ -133,18 +121,6 @@ namespace Tizen.NUI.BaseComponents
         /// </exception>
         /// <since_tizen> 4 </since_tizen>
         public string TranslatablePlaceholderText
-        {
-            get
-            {
-                return GetValue(TranslatablePlaceholderTextProperty) as string;
-            }
-            set
-            {
-                SetValue(TranslatablePlaceholderTextProperty, value);
-            }
-        }
-
-        private string InternalTranslatablePlaceholderText
         {
             get
             {
@@ -1441,18 +1417,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return (bool)GetValue(EnableEditingProperty);
-            }
-            set
-            {
-                SetValue(EnableEditingProperty, value);
-            }
-        }
-
-        private bool InternalEnableEditing
-        {
-            get
-            {
                 bool temp;
                 GetProperty(TextEditor.Property.EnableEditing).Get(out temp);
                 return temp;
@@ -1469,18 +1433,6 @@ namespace Tizen.NUI.BaseComponents
         /// </summary>
         [EditorBrowsable(EditorBrowsableState.Never)]
         public int HorizontalScrollPosition
-        {
-            get
-            {
-                return (int)GetValue(HorizontalScrollPositionProperty);
-            }
-            set
-            {
-                SetValue(HorizontalScrollPositionProperty, value);
-            }
-        }
-
-        private int InternalHorizontalScrollPosition
         {
             get
             {
@@ -1509,18 +1461,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return (int)GetValue(VerticalScrollPositionProperty);
-            }
-            set
-            {
-                SetValue(VerticalScrollPositionProperty, value);
-            }
-        }
-
-        private int InternalVerticalScrollPosition
-        {
-            get
-            {
                 int temp;
                 using (PropertyValue propertyValue = GetProperty(TextEditor.Property.VerticalScrollPosition))
                 {
@@ -1543,18 +1483,6 @@ namespace Tizen.NUI.BaseComponents
         /// </summary>
         [EditorBrowsable(EditorBrowsableState.Never)]
         public int PrimaryCursorPosition
-        {
-            get
-            {
-                return (int)GetValue(PrimaryCursorPositionProperty);
-            }
-            set
-            {
-                SetValue(PrimaryCursorPositionProperty, value);
-            }
-        }
-
-        private int InternalPrimaryCursorPosition
         {
             get
             {
