+++ src/Tizen.NUI/src/public/BaseComponents/TextEditorBindableProperty.cs
@@ -1092,113 +1092,5 @@ namespace Tizen.NUI.BaseComponents
             Tizen.NUI.Object.GetProperty((System.Runtime.InteropServices.HandleRef)textEditor.SwigCPtr, TextEditor.Property.MinLineSize).Get(out temp);
             return temp;
         }));
-
-        /// <summary>
-        /// TranslatableTextProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty TranslatableTextProperty = BindableProperty.Create(nameof(TranslatableText), typeof(string), typeof(Tizen.NUI.BaseComponents.TextEditor), string.Empty, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            if (newValue != null)
-            {
-                instance.InternalTranslatableText = (string)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            return instance.InternalTranslatableText;
-        });
-
-        /// <summary>
-        /// TranslatablePlaceholderTextProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty TranslatablePlaceholderTextProperty = BindableProperty.Create(nameof(TranslatablePlaceholderText), typeof(string), typeof(Tizen.NUI.BaseComponents.TextEditor), string.Empty, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            if (newValue != null)
-            {
-                instance.InternalTranslatablePlaceholderText = (string)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            return instance.InternalTranslatablePlaceholderText;
-        });
-
-        /// <summary>
-        /// EnableEditingProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty EnableEditingProperty = BindableProperty.Create(nameof(EnableEditing), typeof(bool), typeof(Tizen.NUI.BaseComponents.TextEditor), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            if (newValue != null)
-            {
-                instance.InternalEnableEditing = (bool)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            return instance.InternalEnableEditing;
-        });
-
-        /// <summary>
-        /// HorizontalScrollPositionProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty HorizontalScrollPositionProperty = BindableProperty.Create(nameof(HorizontalScrollPosition), typeof(int), typeof(Tizen.NUI.BaseComponents.TextEditor), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            if (newValue != null)
-            {
-                instance.InternalHorizontalScrollPosition = (int)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            return instance.InternalHorizontalScrollPosition;
-        });
-
-        /// <summary>
-        /// VerticalScrollPositionProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty VerticalScrollPositionProperty = BindableProperty.Create(nameof(VerticalScrollPosition), typeof(int), typeof(Tizen.NUI.BaseComponents.TextEditor), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            if (newValue != null)
-            {
-                instance.InternalVerticalScrollPosition = (int)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            return instance.InternalVerticalScrollPosition;
-        });
-
-        /// <summary>
-        /// PrimaryCursorPositionProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty PrimaryCursorPositionProperty = BindableProperty.Create(nameof(PrimaryCursorPosition), typeof(int), typeof(Tizen.NUI.BaseComponents.TextEditor), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            if (newValue != null)
-            {
-                instance.InternalPrimaryCursorPosition = (int)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.TextEditor)bindable;
-            return instance.InternalPrimaryCursorPosition;
-        });
     }
 }
