+++ src/Tizen.NUI/src/public/BaseComponents/TextLabel.cs
@@ -420,18 +420,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return GetValue(ShadowOffsetProperty) as Vector2;
-            }
-            set
-            {
-                SetValue(ShadowOffsetProperty, value);
-            }
-        }
-
-        private Vector2 InternalShadowOffset
-        {
-            get
-            {
                 Vector2 shadowOffset = new Vector2();
                 Shadow.Find(TextLabel.Property.SHADOW, "offset")?.Get(shadowOffset);
                 return new Vector2(OnShadowOffsetChanged, shadowOffset.X, shadowOffset.Y);
@@ -463,18 +451,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return GetValue(ShadowColorProperty) as Vector4;
-            }
-            set
-            {
-                SetValue(ShadowColorProperty, value);
-            }
-        }
-
-        private Vector4 InternalShadowColor
-        {
-            get
-            {
                 Vector4 shadowColor = new Vector4();
                 Shadow.Find(TextLabel.Property.SHADOW, "color")?.Get(shadowColor);
                 return new Vector4(OnShadowColorChanged, shadowColor.X, shadowColor.Y, shadowColor.Z, shadowColor.W);
@@ -502,18 +478,6 @@ namespace Tizen.NUI.BaseComponents
         /// </remarks>
         [Obsolete("Please do not use this UnderlineEnabled(Deprecated). Please use Underline instead.")]
         public bool UnderlineEnabled
-        {
-            get
-            {
-                return (bool)GetValue(UnderlineEnabledProperty);
-            }
-            set
-            {
-                SetValue(UnderlineEnabledProperty, value);
-            }
-        }
-
-        private bool InternalUnderlineEnabled
         {
             get
             {
@@ -549,18 +513,6 @@ namespace Tizen.NUI.BaseComponents
         {
             get
             {
-                return GetValue(UnderlineColorProperty) as Vector4;
-            }
-            set
-            {
-                SetValue(UnderlineColorProperty, value);
-            }
-        }
-
-        private Vector4 InternalUnderlineColor
-        {
-            get
-            {
                 Vector4 underlineColor = new Vector4();
                 Underline.Find(TextLabel.Property.UNDERLINE, "color")?.Get(underlineColor);
                 return new Vector4(OnUnderlineColorChanged, underlineColor.X, underlineColor.Y, underlineColor.Z, underlineColor.W);
@@ -588,18 +540,6 @@ namespace Tizen.NUI.BaseComponents
         /// </remarks>
         [Obsolete("Please do not use this UnderlineHeight(Deprecated). Please use Underline instead.")]
         public float UnderlineHeight
-        {
-            get
-            {
-                return (float)GetValue(UnderlineHeightProperty);
-            }
-            set
-            {
-                SetValue(UnderlineHeightProperty, value);
-            }
-        }
-
-        private float InternalUnderlineHeight
         {
             get
             {
