+++ src/Tizen.NUI/src/public/BaseComponents/ViewBindableProperty.cs
@@ -1836,7 +1836,7 @@ namespace Tizen.NUI.BaseComponents
         /// AccessibilityRoleProperty
         /// </summary>
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty AccessibilityRoleProperty = BindableProperty.Create(nameof(AccessibilityRole), typeof(Role), typeof(View), default(int), propertyChanged: (bindable, oldValue, newValue) =>
+        public static readonly BindableProperty AccessibilityRoleProperty = BindableProperty.Create(nameof(AccessibilityRole), typeof(int), typeof(View), default(int), propertyChanged: (bindable, oldValue, newValue) =>
         {
             var view = (View)bindable;
             if (newValue != null)
@@ -1850,7 +1850,7 @@ namespace Tizen.NUI.BaseComponents
 
             int temp = 0;
             Tizen.NUI.Object.GetProperty((System.Runtime.InteropServices.HandleRef)view.SwigCPtr, View.Property.AccessibilityRole).Get(out temp);
-            return (Role)temp;
+            return temp;
         });
 
         /// <summary>
@@ -1871,312 +1871,6 @@ namespace Tizen.NUI.BaseComponents
             bool temp = false;
             Tizen.NUI.Object.GetProperty((System.Runtime.InteropServices.HandleRef)view.SwigCPtr, View.Property.AccessibilityHighlightable).Get(out temp);
             return temp;
-        });
-
-        /// <summary>
-        /// ExcludeLayoutingProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty ExcludeLayoutingProperty = BindableProperty.Create(nameof(ExcludeLayouting), typeof(bool), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalExcludeLayouting = (bool)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalExcludeLayouting;
-        });
-
-        /// <summary>
-        /// TooltipTextProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty TooltipTextProperty = BindableProperty.Create(nameof(TooltipText), typeof(string), typeof(Tizen.NUI.BaseComponents.View), string.Empty, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalTooltipText = (string)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalTooltipText;
-        });
-
-        /// <summary>
-        /// PositionUsesAnchorPointProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty PositionUsesAnchorPointProperty = BindableProperty.Create(nameof(PositionUsesAnchorPoint), typeof(bool), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalPositionUsesAnchorPoint = (bool)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalPositionUsesAnchorPoint;
-        });
-
-        /// <summary>
-        /// AnchorPointProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty AnchorPointProperty = BindableProperty.Create(nameof(AnchorPoint), typeof(Tizen.NUI.Position), typeof(Tizen.NUI.BaseComponents.View), null, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalAnchorPoint = (Tizen.NUI.Position)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalAnchorPoint;
-        });
-
-        /// <summary>
-        /// WidthSpecificationProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty WidthSpecificationProperty = BindableProperty.Create(nameof(WidthSpecification), typeof(int), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalWidthSpecification = (int)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalWidthSpecification;
-        });
-
-        /// <summary>
-        /// HeightSpecificationProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty HeightSpecificationProperty = BindableProperty.Create(nameof(HeightSpecification), typeof(int), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalHeightSpecification = (int)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalHeightSpecification;
-        });
-
-        /// <summary>
-        /// LayoutTransitionProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty LayoutTransitionProperty = BindableProperty.Create(nameof(LayoutTransition), typeof(Tizen.NUI.LayoutTransition), typeof(Tizen.NUI.BaseComponents.View), null, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalLayoutTransition = (Tizen.NUI.LayoutTransition)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalLayoutTransition;
-        });
-
-        /// <summary>
-        /// PaddingEXProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty PaddingEXProperty = BindableProperty.Create(nameof(PaddingEX), typeof(Tizen.NUI.Extents), typeof(Tizen.NUI.BaseComponents.View), null, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalPaddingEX = (Tizen.NUI.Extents)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalPaddingEX;
-        });
-
-        /// <summary>
-        /// LayoutProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty LayoutProperty = BindableProperty.Create(nameof(Layout), typeof(Tizen.NUI.LayoutItem), typeof(Tizen.NUI.BaseComponents.View), null, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalLayout = (Tizen.NUI.LayoutItem)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalLayout;
-        });
-
-        /// <summary>
-        /// BackgroundImageSynchronosLoadingProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty BackgroundImageSynchronosLoadingProperty = BindableProperty.Create(nameof(BackgroundImageSynchronosLoading), typeof(bool), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalBackgroundImageSynchronosLoading = (bool)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalBackgroundImageSynchronosLoading;
-        });
-
-        /// <summary>
-        /// BackgroundImageSynchronousLoadingProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty BackgroundImageSynchronousLoadingProperty = BindableProperty.Create(nameof(BackgroundImageSynchronousLoading), typeof(bool), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalBackgroundImageSynchronousLoading = (bool)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalBackgroundImageSynchronousLoading;
-        });
-
-        /// <summary>
-        /// EnableControlStatePropagationProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty EnableControlStatePropagationProperty = BindableProperty.Create(nameof(EnableControlStatePropagation), typeof(bool), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalEnableControlStatePropagation = (bool)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalEnableControlStatePropagation;
-        });
-
-        /// <summary>
-        /// GrabTouchAfterLeaveProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty GrabTouchAfterLeaveProperty = BindableProperty.Create(nameof(GrabTouchAfterLeave), typeof(bool), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalGrabTouchAfterLeave = (bool)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalGrabTouchAfterLeave;
-        });
-
-        /// <summary>
-        /// BlendEquationProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty BlendEquationProperty = BindableProperty.Create(nameof(BlendEquation), typeof(Tizen.NUI.BlendEquationType), typeof(Tizen.NUI.BaseComponents.View), null, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalBlendEquation = (Tizen.NUI.BlendEquationType)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalBlendEquation;
-        });
-
-        /// <summary>
-        /// TransitionOptionsProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty TransitionOptionsProperty = BindableProperty.Create(nameof(TransitionOptions), typeof(Tizen.NUI.TransitionOptions), typeof(Tizen.NUI.BaseComponents.View), null, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalTransitionOptions = (Tizen.NUI.TransitionOptions)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalTransitionOptions;
-        });
-
-        /// <summary>
-        /// AutomationIdProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty AutomationIdProperty = BindableProperty.Create(nameof(AutomationId), typeof(string), typeof(Tizen.NUI.BaseComponents.View), string.Empty, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalAutomationId = (string)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalAutomationId;
-        });
-
-        /// <summary>
-        /// TouchAreaOffsetProperty
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static readonly BindableProperty TouchAreaOffsetProperty = BindableProperty.Create(nameof(TouchAreaOffset), typeof(Tizen.NUI.Offset), typeof(Tizen.NUI.BaseComponents.View), 0, propertyChanged: (bindable, oldValue, newValue) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            if (newValue != null)
-            {
-                instance.InternalTouchAreaOffset = (Tizen.NUI.Offset)newValue;
-            }
-        },
-        defaultValueCreator: (bindable) =>
-        {
-            var instance = (Tizen.NUI.BaseComponents.View)bindable;
-            return instance.InternalTouchAreaOffset;
         });
 
         private void SetBackgroundImage(string value)
